Update entrypoint to add a rerun # indicator
Reruns are when a task executes again in the game tfo resource generation. This happens in the event of pod restarts. Logs are captured as brand new for reruns.

- added new tftask builder
- added setup script automation
- replaced old entrypoint shell script with c++ executable

The rerun will become important when approval workflow is introduced. Approval workflows will pause and wait for approval from a user to continue with the execution of the terraform plan. The approval must match the rerun in order to continue (hypothetically... at least thats the plan).